Unpack into make_map instead of using to_map
This fixes the code generation issues. The problem is that `to_map` allows for
duplicate keys in the "from" sequence, which means it needs to recursively
insert into a `hana::map`. This is both very bad for codegen and for
compile-times. `make_map`, on the other hand, requires all the keys to be
distinct, which makes it much more straightforward and much easier to optimize.